유효성체크 클래스 생성
