+++ src/lotto/view/Input.java
@@ -1,10 +1,8 @@
 package lotto.view;
 
 import java.util.Arrays;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Scanner;
-import java.util.Set;
 import java.util.stream.Collectors;
 
 public class Input {
@@ -15,12 +13,6 @@ public class Input {
     private static final String BONUS_NUMBER_INFO = "보너스 번호를 입력해주세요.";
     private static final String REQUEST_LOTTO_NUMBER_INFO = "수동으로 구매할 로또 수를 입력해 주세요.";
     private static final String REQUEST_LOTTO_NUMBERS_INFO = "수동으로 구매할 번호를 입력해 주세요.";
-    private static final int LOTTO_NUMBERS_LENGTH = 6;
-    private static final String ERROR_MESSAGE_FOR_LOTTO_NUMBERS_LENGTH = "로또 번호는 6개를 입력해주세요.";
-    private static final int MIN_VALUE = 1;
-    private static final int MAX_VALUE = 45;
-    private static final String ERROR_MESSAGE_FOR_LOTTO_NUMBER_RANGE = "로또번호는 1~45 범위의 숫자로 입력해주세요.";
-    private static final String ERROR_MESSAGE_FOR_LOTTO_NUMBER_DUPLICATION = "로또번호는 서로 다른 숫자로 입력해주세요.";
 
     private Input() {
     }
@@ -46,7 +38,7 @@ public class Input {
             .boxed()
             .collect(Collectors.toList());
 
-        return isValidLottoNumbers(lottoNumbers);
+        return lottoNumbers;
     }
 
     public static List<Integer> getLuckyNumbers() {
@@ -64,22 +56,5 @@ public class Input {
     public static void getLottoNumbersInfo() {
         System.out.println(REQUEST_LOTTO_NUMBERS_INFO);
 
-    }
-
-
-    private static List<Integer> isValidLottoNumbers(List<Integer> lottoNumbers) {
-        if (lottoNumbers.size() != LOTTO_NUMBERS_LENGTH) {
-            throw new IllegalArgumentException(ERROR_MESSAGE_FOR_LOTTO_NUMBERS_LENGTH);
-        }
-        boolean isNotValidRangeLottoNumber = lottoNumbers.stream()
-            .anyMatch((number) -> (number < MIN_VALUE || number > MAX_VALUE));
-        if (isNotValidRangeLottoNumber) {
-            throw new IllegalArgumentException(ERROR_MESSAGE_FOR_LOTTO_NUMBER_RANGE);
-        }
-        Set<Integer> abcd = new HashSet<>(lottoNumbers);
-        if (abcd.size() != LOTTO_NUMBERS_LENGTH) {
-            throw new IllegalArgumentException(ERROR_MESSAGE_FOR_LOTTO_NUMBER_DUPLICATION);
-        }
-        return lottoNumbers;
     }
 }
